Bringing back `.jsx` extension for JSX files
