+++ components/common/Canvas.jsx
