+++ components/screens/Home.jsx
@@ -1,13 +1,14 @@
 import React, { useContext, useState, useMemo } from "react";
 import { ActivityIndicator, StyleSheet, View, PixelRatio, FlatList } from "react-native";
 import { Divider, Layout, Text } from "@ui-kitten/components";
-import { Carousel, Polyptych, Mark, Shelf, Wildcard } from "../";
-import { useWordPress } from "../../hooks/useWordPress";
-import { Sections, Spacing } from "../../utils/constants";
 import _ from "lodash";
 import { DeviceType } from "expo-device";
-import { ThemeContext } from "../../theme-context";
 import { StatusBar } from "expo-status-bar";
+
+import { Carousel, Polyptych, Mark, Shelf, Wildcard } from "..";
+import { useWordPress } from "../../hooks/useWordPress";
+import { Sections, Spacing } from "../../utils/constants";
+import { ThemeContext } from "../../theme-context";
 
 const FirstPage = React.memo(({ articles, data, navigation, deviceType }) => {
   const groupSize = deviceType === DeviceType.PHONE ? 1 : 2;
@@ -15,29 +16,29 @@ const FirstPage = React.memo(({ articles, data, navigation, deviceType }) => {
   return (
     <React.Fragment>
       <Carousel articles={articles[Sections.FEATURED.slug] ?? []} navigation={navigation} />
-      <Mark category={Sections.NEWS} seed={data[Sections.NEWS.slug] ?? []} navigation={navigation} />
+      <Mark category={Sections.NEWS} navigation={navigation} seed={data[Sections.NEWS.slug] ?? []} />
       <Polyptych articles={articles[Sections.NEWS.slug] ?? []} navigation={navigation} />
       <Divider marginTop={Spacing.medium} />
-      <Mark category={Sections.OPINIONS} seed={data[Sections.OPINIONS.slug]} navigation={navigation} />
+      <Mark category={Sections.OPINIONS} navigation={navigation} seed={data[Sections.OPINIONS.slug]} />
       <Shelf articles={articles[Sections.OPINIONS.slug] ?? []} navigation={navigation} />
-      <Mark category={Sections.SPORTS} seed={data[Sections.SPORTS.slug] ?? []} navigation={navigation} />
+      <Mark category={Sections.SPORTS} navigation={navigation} seed={data[Sections.SPORTS.slug] ?? []} />
       <Polyptych articles={articles[Sections.SPORTS.slug] ?? []} navigation={navigation} />
       <Divider marginTop={Spacing.medium} />
       {_.chunk(articles?.culture?.slice(0, 4 * groupSize), groupSize)?.map((group, outerIndex) => (
-        <View style={{ flex: 1 / groupSize, flexDirection: "row" }} key={outerIndex}>
+        <View key={outerIndex} style={{ flex: 1 / groupSize, flexDirection: "row" }}>
           {group.map((item, index) => (
             <Wildcard
-              item={item}
-              index={outerIndex * index + index}
               key={item.id.toString()}
+              index={outerIndex * index + index}
+              item={item}
               navigation={navigation}
             />
           ))}
         </View>
       ))}
       <Divider />
-      <Mark category={Sections.HUMOR} seed={data[Sections.HUMOR.slug] ?? []} alternate navigation={navigation} />
-      <Shelf articles={articles[Sections.HUMOR.slug] ?? []} alternate navigation={navigation} />
+      <Mark alternate category={Sections.HUMOR} navigation={navigation} seed={data[Sections.HUMOR.slug] ?? []} />
+      <Shelf alternate articles={articles[Sections.HUMOR.slug] ?? []} navigation={navigation} />
       <Divider />
     </React.Fragment>
   );
@@ -50,36 +51,34 @@ export default function Home({ navigation, route }) {
   // Ensure that no culture posts simultaneously appear in wildcard.
   const wildcard = data?.wildcard?.filter(item => !articles?.culture?.find(article => article.id === item.id));
 
-  const Header = useMemo(() => {
-    return (
-      <FirstPage
-        articles={articles}
-        data={data}
-        navigation={navigation}
-        deviceType={deviceType}
-      />
-    );
-  }, [articles, data, navigation, deviceType]);
+  const Header = useMemo(() => (
+    <FirstPage
+      articles={articles}
+      data={data}
+      deviceType={deviceType}
+      navigation={navigation}
+    />
+  ), [articles, data, navigation, deviceType]);
 
   return articles ? (
     <Layout style={styles.container}>
       <StatusBar style={theme === "dark" ? "light" : "dark"} />
       <FlatList
-        ListHeaderComponent={Header}
         data={wildcard}
+        keyExtractor={item => item.id.toString()}
+        ListFooterComponent={() => <ActivityIndicator style={{ marginBottom: Spacing.large }} />}
+        ListHeaderComponent={Header}
+        onEndReached={() => setPageNumber(pageNumber + 1)}
+        onEndReachedThreshold={0.25}
         renderItem={(post, index) => (
           <Wildcard
-            item={post.item}
-            index={index}
             key={post.item.id.toString()}
+            index={index}
+            item={post.item}
             navigation={navigation}
             verbose
           />
         )}
-        onEndReached={() => setPageNumber(pageNumber + 1)}
-        onEndReachedThreshold={0.25}
-        keyExtractor={item => item.id.toString()}
-        ListFooterComponent={() => <ActivityIndicator style={{ marginBottom: Spacing.large }} />}
       />
     </Layout>
   ) : (
@@ -88,7 +87,7 @@ export default function Home({ navigation, route }) {
     </Layout>
   );
 }
-  
+
 const styles = StyleSheet.create({
   container: {
     flex: 1
