+++ hooks/useWordPress.js
@@ -1,20 +1,20 @@
-import { useState, useEffect } from "react"
-import { Sections } from "../utils/constants"
-import Model from "../utils/model"
-import _ from "lodash"
+import { useState, useEffect } from "react";
+import { Sections } from "../utils/constants";
+import Model from "../utils/model";
+import _ from "lodash";
 
-const BATCH_SIZE = 6
-const PAGE_SIZE = 12
-const homeCount = BATCH_SIZE * Object.keys(Sections).length
+const BATCH_SIZE = 6;
+const PAGE_SIZE = 12;
+const homeCount = BATCH_SIZE * Object.keys(Sections).length;
 
 export const useWordPress = (pageNumber = 1) => {
-    const [data, setData] = useState({})
-    const [articles, setArticles] = useState({})
-    const [loading, setLoading] = useState(true)
-    const [error, setError] = useState(null)
+    const [data, setData] = useState({});
+    const [articles, setArticles] = useState({});
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(null);
   
     useEffect(() => {
-      setLoading(true)
+      setLoading(true);
   
       Object.values(Sections).forEach(async category => {
         try {
@@ -22,27 +22,27 @@ export const useWordPress = (pageNumber = 1) => {
           setData((prevState) => ({
             ...prevState,
             [category.slug]: [...(category.slug in prevState ? prevState[category.slug] : []), ...response]
-          }))
+          }));
           
           if (pageNumber === 1) {
             setArticles((prevState) => ({
               ...prevState,
               [category.slug]: response.filter(item => !item.categories.includes(Sections.FEATURED.id) || category.slug === Sections.FEATURED.slug)
-            }))
+            }));
 
             if (category.slug === Sections.ARTS_LIFE.slug || category.slug === Sections.THE_GRIND.slug) {
               setArticles((prevState) => ({
                 ...prevState,
                 culture: [...(prevState?.culture ?? []), ..._.shuffle(response)]
-              }))
+              }));
             }
           }
         } catch (error) {
-          setError(error)
+          setError(error);
         } finally {
-          setLoading(false)
+          setLoading(false);
         }
-      })
+      });
 
       Model.posts().perPage(PAGE_SIZE).page(pageNumber + Math.ceil(homeCount / PAGE_SIZE) + 1).get().then(posts => {
         setData((prevState) => ({
@@ -51,9 +51,9 @@ export const useWordPress = (pageNumber = 1) => {
             ...(prevState?.wildcard ?? []),
             ...posts.filter(post => !Object.keys(articles).some(category => articles[category].some(item => item.id === post.id)))
           ]
-        }))
-      })
-    }, [pageNumber])
+        }));
+      });
+    }, [pageNumber]);
     
-    return { data, articles, loading, error }
-}
+    return { data, articles, loading, error };
+};
